legacy_streamlit/requirements.txt
streamlit
requests
python-dotenv

# uv pip install -r requirements.txt